requirements.txt
asttokens==2.2.1
backcall==0.2.0
brotlipy==0.7.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
comm==0.1.2
cryptography @ file:///croot/cryptography_1665612644927/work
debugpy==1.6.4
decorator==4.4.2
dgl==0.9.1.post1
entrypoints==0.4
executing==1.2.0
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fst-pso==1.8.1
FuzzyTM==2.0.5
gensim==4.3.0
idna @ file:///croot/idna_1666125576474/work
ipykernel==6.19.2
ipython==8.7.0
ipywidgets==8.0.3
jedi==0.18.2
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib @ file:///croot/joblib_1666298844297/work
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyterlab-widgets==3.0.4
karateclub==1.3.3
Levenshtein==0.20.9
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib-inline==0.1.6
miniful==0.0.6
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
nest-asyncio==1.5.6
networkx==2.6.3
numpy==1.22.4
packaging==22.0
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.0
prompt-toolkit==3.0.36
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyFUME==0.2.25
Pygments==2.13.0
PyGSP==0.5.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-Levenshtein==0.20.9
python-louvain==0.16
pytz==2022.7
pyzmq==24.0.1
rapidfuzz==2.13.7
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
scikit-learn==1.1.1
scipy==1.9.3
simpful==2.9.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.3.0
stack-data==0.6.2
threadpoolctl==3.1.0
torch==1.12.0
torch-cluster @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-cluster_1656602642185/work
torch-geometric @ file:///usr/share/miniconda/envs/test/conda-bld/pyg_1669878904337/work
torch-scatter @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-scatter_1669298464990/work
torch-sparse @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-sparse_1665388439494/work
tornado==6.2
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
traitlets==5.8.0
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
urllib3 @ file:///croot/urllib3_1670526988650/work
wcwidth==0.2.5
widgetsnbextension==4.0.4
